IO port framework now works
A68k is defaulting all my address moves to word size, rather than
long.  I don't know whether I've expunged all this nonsense yet, but
I'm trying.

This version executes an infinite loop which writes 'A' to port 00h.
I've patched in a write-to-console function on port 00h, so this can
be used as a sort of debug monitor.