Added guards around all routines that touch ePC or eSP.
These guards have one major fault I can see.  I put them as early in
the routine as possible, but it's still a distinct possibility that
the 68k interrupt will fire between

    move.b  (epc)+,d0

in macro DONE of one instruction and the call to HOLD_INTS in the
following instruction.  I don't have a good solution to this.  I can
use the hardware interrupt holding support to make everything a
critical section except for the cycle gap before that instruction, but
that makes _every_ instruction 24 cycles slower.  I don't consider
that an acceptable solution.