Non-code changes (comments and macro arguments)
I added a fourth and fifth argument to OPCODE().  Now it accepts both
of the arguments for the TIME macro (to allow for more precise speed
emulation in the distant future) as well as a fifth argument, the size
of the opcode routine (excluding START and DONE).  I think this will
allow me to automatically generate minimal offsets when I finally use
Tezxas's dispatch routine.